requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-google-genai==1.0.10
python-dotenv==1.0.1
pydantic==2.8.2
pytest==8.3.2
httpx==0.27.2
requests==2.32.3
google-search-results==2.4.2
streamlit==1.39.0
